+++ activesupport/CHANGELOG.md
@@ -3,11 +3,16 @@
 
     To opt-in load the [listen](https://github.com/guard/listen) gem in `Gemfile`:
 
-      group :development do
-        gem 'listen', '~> 3.0.4'
-      end
+        group :development do
+          gem 'listen', '~> 3.0.4'
+        end
 
     *Puneet Agarwal* and *Xavier Noria*
+
+* Added `Time#days_in_year` to return the number of days in the given year, or the
+  current year if no argument is provided.
+
+  *Jon Pascoe*
 
 * Updated `parameterize` to preserve the case of a string, optionally.